rename transaction charts
